refactor(api): update enums and improve component rendering
- Added new user position and status enums for better clarity and type safety.
- Updated labels in TeamStatusData for consistency.
- Refactored TeamDetailModal and UserDetailModal to utilize new enum structures for rendering statuses and roles.
- Enhanced type safety by ensuring correct usage of enums in various components.
