+++ src/api/enum.ts
@@ -755,21 +755,21 @@ export enum TeamStatus {
  * UserPosition
  */
 export enum UserPosition {
-  USER_POSITION_ADMIN,
-  USER_POSITION_GUEST,
-  USER_POSITION_SUPER_ADMIN,
   USER_POSITION_UNKNOWN,
-  USER_POSITION_USER
+  USER_POSITION_SUPER_ADMIN,
+  USER_POSITION_ADMIN,
+  USER_POSITION_USER,
+  USER_POSITION_GUEST,
 }
 
 /**
  * UserStatus
  */
 export enum UserStatus {
-  USER_STATUS_DELETED,
-  USER_STATUS_FORBIDDEN,
+  USER_STATUS_UNKNOWN,
   USER_STATUS_NORMAL,
-  USER_STATUS_UNKNOWN
+  USER_STATUS_FORBIDDEN,
+  USER_STATUS_DELETED,
 }
