+++ src/pages/system/user/modal-detail.tsx
@@ -1,7 +1,7 @@
-import { UserItem } from '@/api/common.types'
-import { UserPosition } from '@/api/enum'
-import { UserStatusData } from '@/api/global'
+import { UserPosition, UserStatus } from '@/api/enum'
+import { UserPositionData, UserStatusData } from '@/api/global'
 import { getUser } from '@/api/request/system'
+import { UserItem } from '@/api/request/types'
 import { useRequest } from 'ahooks'
 import { Avatar, Badge, Descriptions, type DescriptionsProps, Modal } from 'antd'
 import { useEffect, useState } from 'react'
@@ -38,14 +38,17 @@ export function DetailModal(props: UserDetailModalProps) {
         <div className='flex items-center gap-2'>
           <Avatar src={detail?.avatar}>{detail?.nickname || detail?.username}</Avatar>
           {detail?.nickname || detail?.username}
-          <Badge color={UserStatusData[detail?.status]?.color} />
+          <Badge color={(detail?.status != null ? UserStatusData[detail.status as UserStatus]?.color : undefined) || 'default'} />
         </div>
       ),
       span: { xs: 1, sm: 2, md: 3, lg: 3, xl: 2, xxl: 2 }
     },
     {
       label: '角色',
-      children: UserPosition[detail?.position],
+      children: (() => {
+        const data = detail?.position != null ? UserPositionData[detail.position as UserPosition] : undefined
+        return <span style={{ color: data?.color || 'gray' }}>{(data as any)?.text ?? data?.label ?? '未知'}</span>
+      })(),
       span: { xs: 1, sm: 2, md: 3, lg: 3, xl: 2, xxl: 2 }
     },
